refactor: creates base abstract class for thumbnails scraper
